Move test-properties out from Device
