+++ miio/cli.py
@@ -12,6 +12,7 @@ from miio.click_common import (
 from miio.miioprotocol import MiIOProtocol
 
 from .cloud import cloud
+from .devtools import devtools
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -60,6 +61,7 @@ def discover(mdns, handshake, network, timeout):
 
 cli.add_command(discover)
 cli.add_command(cloud)
+cli.add_command(devtools)
 
 
 def create_cli():
